+++ ai/heuristics/heuristics.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   heuristics.hpp                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mmarinel <mmarinel@student.42.fr>          +#+  +:+       +#+        */
+/*   By: cy4gate_mmarinelli <cy4gate_mmarinelli@    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 13:01:18 by matteo            #+#    #+#             */
-/*   Updated: 2024/06/02 20:04:42 by mmarinel         ###   ########.fr       */
+/*   Updated: 2024/06/05 10:18:51 by cy4gate_mma      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,9 +43,9 @@ hfromString = {
 class	t_Iordering_func
 {
 public:
-	virtual uint8_t		f_val(const Node* n) const = 0;
-	virtual uint8_t		h(const Node* n) const = 0;
-	virtual uint8_t		g(const Node* n) const = 0;
+	virtual int		f_val(const Node* n) const = 0;
+	virtual int		h(const Node* n) const = 0;
+	virtual int		g(const Node* n) const = 0;
 	virtual bool		cmp(const Node* n1, const Node* n2) const = 0;
 };
 
@@ -53,17 +53,17 @@ template <typename H>
 class t_RbsOrdering_func: public t_Iordering_func
 {
 public:
-	virtual uint8_t			f_val(const Node* n) const override {
+	virtual int			f_val(const Node* n) const override {
 		if (-1 == n->f)
 			const_cast<Node*>(n)->f = g(n) + H::getInstance()(n);
 		return n->f;
 	}
 
-	virtual uint8_t	h(const Node* n) const override {
+	virtual int	h(const Node* n) const override {
 		return H::getInstance()(n);
 	}
 
-	virtual uint8_t	g(const Node* n) const override {
+	virtual int	g(const Node* n) const override {
 		return n->pCost;
 	}
 	
@@ -77,7 +77,7 @@ public:
 class t_manhattan_score: public t_RbsOrdering_func<t_manhattan_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 	
 	static const t_manhattan_score&	getInstance();
 private:
@@ -87,7 +87,7 @@ private:
 class t_linear_conflict_score: public t_RbsOrdering_func<t_linear_conflict_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 	
 	static const t_linear_conflict_score&	getInstance();
 private:
@@ -97,7 +97,7 @@ private:
 class t_corner_tiles_score: public t_RbsOrdering_func<t_corner_tiles_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 	
 	static const t_corner_tiles_score&	getInstance();
 private:
@@ -108,7 +108,7 @@ class t_misplaced_tiles_score:
 	public t_RbsOrdering_func<t_misplaced_tiles_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 	
 	static const t_misplaced_tiles_score&	getInstance();
 private:
@@ -118,7 +118,7 @@ private:
 class t_gaschnig_score: public t_RbsOrdering_func<t_gaschnig_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 
 	static const t_gaschnig_score&	getInstance();
 private:
@@ -128,7 +128,7 @@ private:
 class t_coalesce_score: public t_RbsOrdering_func<t_coalesce_score>
 {
 public:
-	uint8_t	operator() (const Node* n) const;
+	int	operator() (const Node* n) const;
 
 	static const t_coalesce_score& getInstance();
 private:
